python-service/pyproject.toml
[project]
name = "python-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery==5.4.0",
    "fastapi==0.115.12",
    "gevent==24.11.1",
    "google-genai==1.7.0",
    "langchain==0.3.23",
    "langchain-core==0.3.51",
    "langchain-google-genai==2.1.2",
    "langgraph==0.3.25",
    "moviepy==2.1.2",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "pymupdf==1.25.5",
    "python-docx==1.1.2",
    "python-multipart==0.0.20",
    "pytubefix==8.12.2",
    "qdrant-client==1.13.3",
    "redis==5.2.1",
    "reportlab==4.4.0",
    "unstructured[all-docs]==0.17.2",
    "uvicorn==0.34.0",
]
